+++ testResources/template-script.lua
@@ -1,0 +1,19 @@
+--declare variables above all the functions i think
+
+function onUse(state,world,pos,player,hand,hit)
+    for key in pairs(state) do
+        print(key)
+        -- make sure you launch minecraft with the console
+        -- enabled, so you can see this (this will be fixed later)
+    end
+end
+
+function onBroken(world,pos,state)
+
+end
+function onSteppedOn(world,pos,state,entity)
+
+end
+function onPlaced(world,pos,state,placer,itemStack)
+
+end
